(maint) Update Travis CI Ruby versions
Prior to this commit preview was running in Travis CI with three
Ruby versions that are all no longer supported by Puppet. Update
the .travis.yml file to use more current versions. Soon this will
be module sync'd and this file will be managed that way, but in
the mean time this will help unblock work for preview.